handlers: Change parameters to allow passing both data and scriptfn
Handlers are normally called with their opaque data pointer as the
second argument.  However, for script handlers the second argument is
a pointer to the script_fn (PREINSTALL or POSTINSTALL) indicating
which phase the handler is being called for.

The Lua handler wrapper code stores a reference to the function as the
handler data, so a handler written in Lua cannot be a script handler.

Change the handler function type to have separate data and scriptfn
parameters, and update all callers and implementations accordingly.

Signed-off-by: Ben Hutchings <ben.hutchings_ext@softathome.com>