Lua: Add support for script handlers
Currently only the image parameter is passed to handlers written in
Lua.  Add the scriptfn parameter, so that it is possible to write
useful script handlers.

This converts the enumerated value to a string value of either
"preinst" or "postinst", for consistency with the way package-
provided scripts are called.

Signed-off-by: Ben Hutchings <ben.hutchings_ext@softathome.com>