Lua: Log exception message for external handler initialisation
If we fail to initialise the swupdate_handlers package, we currently
only log a generic error message plus a list of paths that were
searched.  If the swupdate_handlers package exists but has a syntax or
initialisation error, this does not help to understand the problem.

Do the same result handling as for the embedded case, so the exception
message is logged in both cases.

Signed-off-by: Ben Hutchings <ben.hutchings_ext@softathome.com>