corelib/lua_interface: added 'progress' method
This makes it possible to send messages from the Lua realm (e.g.
handlers) to clients listening on the IPC progress interface.

For example, this could be used to extract some data from the update
package and show it to the user (which could be the duty of the IPC
listener).

Signed-off-by: Christian Storm <christian.storm@siemens.com>
Signed-off-by: Michael Adler <michael.adler@siemens.com>