progress_ipc: do not log to stdout
As a library, progress_ipc should not log to stdout as it may break
the interfaces towards other applications. For example, it may be
required that each line written to stdout is valid JSON.
Of course, there are workarounds for this, e.g. writing a wrapper shell
script which filters the relevant messages. This is ugly though and
somewhat defeats the purpose of using the library: if I use the library,
I want to have full control and not rely on any hacks.

Signed-off-by: Christian Storm <christian.storm@siemens.com>
Signed-off-by: Michael Adler <michael.adler@siemens.com>