scripts: Add checkpatch support
checkpatch is a useful script found on the Linux and U-Boot
sources.

It helps to detect coding style issues, misspelling, etc
on a given patch.

Import the checkpatch support from U-Boot 2022.01 tree with some
discarded changes in .checkpatch.conf that does not apply to swupdate.

Usage to verify a patch:

./scripts/checkpatch.pl 0001-mychange.patch

Usage to verify a file:

./scripts/checkpatch.pl -f myfile.c

Signed-off-by: Fabio Estevam <festevam@gmail.com>
Signed-off-by: Stefano Babic <sbabic@denx.de>