diskpart_handler: support explicitly sized partitions
When passing in a partition size in units of sectors, fdisk expects you to
enable explicit sizing on the partition; otherwise, fdisk assumes you want
the partition aligned to the nearest grain (typically 1MB), which may
corrupt a partition table containing already explicitly sized partitions.

This matches the behavior of other libfdisk clients in util-linux (see
excerpt from libfdisk's parse_line_nameval() below):

libfdisk/src/script.c:
        if (pow)        /* specified as <num><suffix> */
                num /= dp->cxt->sector_size;
        else            /* specified as number of sectors */
                fdisk_partition_size_explicit(pa, 1);

To achieve this, we allow ustrtoull() to return a char* to the suffix
of the converted string, if one exists.  If strtoull() points to a different
suffix in the string than ustrtoull(), we assume ustrtoull() consumed a
user provided size-type delimiter.  If no size-type delimiter match is found,
explicit sizing is not enabled and fdisk aligns the partition to the nearest
grain.

Signed-off-by: Kyle Russell <bkylerussell@gmail.com>