Set ustate when an update is started
ustate is just set after an update with the result, but it was foreseen
to set it to IN_PROGRESS, too, when an update runs. This information is
really srored in the marker (recovery_status bootloader variable).
However, the two variables are not set in an atomic way because they are
used in different use cases (recovery_status in single-copy strategy).

Align ustate to recovery_status, and set it when an update is started. A
power-cut during an update can be detected by examining the value of the
variable and this can be useful in some use cases (reporting a failed
update due to a power-cut).

Factorize code to set the variables in a separate function to make the
code more readabla.

Signed-off-by: Stefano Babic <sbabic@denx.de>