network_ipc async_thread: fix hang on failed update
if a update is large and fails, it's likely write_image would fail.
Old code would just exit there, quitting swupdate but a recent cleanup
made this just exit async_thread without calling the end function, so
the caller would be left hanging waiting for an update that will never
finish.

This can easily be reproduced with a large script that exits immediately
with a failure code with swupdate -i or swupdate_client.

Fixes: 626d83f8819d ("IPC: do not call exit")
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>