install_from_file: initialize and take mutex earlier
if async_thread finishes really fast, it's possible it'd try to call end
callback to signal us before we're waiting for it, leaving us hanging here.

This completes the previous patch for swupdate -i

Mymutex is not a descriptive variable name and has been renamed to
install_file_mutex, as it is barely used this was left in same patch.

Also, making the error case return path shared also fixes 'close(fd)'
that should have been checking for filename being set.

Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>