core/stream_interface.c: increment progress step
Increment the progress step before sending swupdate_progress_end.
On success, this will be needed by swupdate-progress tool to check if
reboot can be done on a successful update.

Currently swupdate_progress_inc_step is only implemented while
installing image from install_single_image. But if image is not specified
in sw-description, and update is done via suricatta or webserver, the progress
step cur_step is never incremented.

If the image is not mentioned in the description file, it makes sense to
send empty string for image handler name and image handler descriptor
to swupdate_progress_inc_step. For the other case, if image is
mentioned, it will hit install_single_image and do the increment of
cur_step in the conventional way.

Signed-off-by: Viraj Shah <viraj.shah@linutronix.de>