run_system_cmd: create additional pipes for INFO and WARN traces
subprocesses (e.g. scripts) are currently limited to ERROR and TRACE.

When TRACE is not displayed (as it can get quite verbose), the only way to
get messages to the user is to write to stderr which will be displayed
in red as ERROR messages.
This can be confusing when we just want to notify users e.g. of upcoming
reboot.

This commits creates a couple of additional pipes which are just handled
like stdout and stderr, except they are displayed at info and warn
log levels.
Scripts can check for SWUPDATE_INFO_FD / SWUPDATE_WARN_FD to decide if
they can output to these file descriptors.

Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>