suricatta/lua: Add suricatta Lua module infrastructure
The server_lua.c C-to-Lua bridge enables writing suricatta modules in Lua.

It provides the infrastructure in terms of the interface to SWUpdate "core"
to the Lua realm, enabling the "business logic" such as handling update
flows and communicating with backend server APIs to be modeled in Lua.

The server_lua.c C-to-Lua bridge provides the same functionality as the
other suricatta modules written in C have, realizing a separation of means
and control. Effectively, it lifts the suricatta interface to the Lua realm.

Signed-off-by: Christian Storm <christian.storm@siemens.com>