handlers: diskpart_handler.c: add install_gpt_partition_image
Add a handler to write in a GPT partition.
With the handler raw, the device of the partition is provided
to the raw handler. This handler use a device on the disk that
contains the GPT partition and the name of the GPT partition
to found the device of the partition to write.

For example:
images: (
	{
		filename = "u-boot.bin";
		type = "gptpart";
		device = "/dev/vdb";
		volume = "u-boot-1";
		offset = "1024";
	},
	{
		filename = "kernel.bin";
		type = "gptpart";
		device = "/dev/vdb";
		volume = "kernel-1";
	},
);

Signed-off-by: Philippe Reynes <philippe.reynes@softathome.com>