Do not call sd_listen_fds() on cleanup
systemd creates the Unix Domain Sockets just once, but duplicates the
fds each time the daemon is started. According to systemd descritpion,
sd_listen_fds() returns the number of passed file descriptors when it is
called for the first time, successive calls return 0. It is then wrong
to call again sd_listen_fds() to decide if the socket must be removed,
because this call will always return 0 and will constraint SWUpdate to
remove the sockets. The effect is that a "systemctl restart swupdate"
will lead to a not working SWUpdate, because systemd will still pass the
file descriptors, but the socket files were removed by previous run.

Signed-off-by: Stefano Babic <sbabic@denx.de>
Reported-by: Marco Braga <marco.braga@gmail.com>