lua: add stat() function
It is sometimes useful to get more information inside a Lua script about
a file. Lua does  not provide direct access, so this is a wrapper for
the stat() function and creates a table that can be parsed by the
caller. Reference for the implementation was luafilesystem (MIT), but
return values were changed to made it more easy to use inside SWUpdate.

Signed-off-by: Stefano Babic <sbabic@denx.de>