parser: allow both string and number for "offset"
The offset attribute can be misleading. It is defined as string and then
converted by the parser, but it is usually set as number in decimal or
hexadecimal format.

In case a number is put in sw-description, the offset is set to 0. This
can cause to write at a wrong offset and no warning or error are raised.

Change this and let to set the offset in both formats. These example
will generate the same offset for SWUpdate:

as number like:

	offset = 1024;
		or
	offset = 0x400;

or as a string:
	offset = "1024";
		or
	offset = "1K";

Signed-off-by: Stefano Babic <sbabic@denx.de>