mongoose: Adapt interface for mongoose 7.8
There are significant differences in the new mongoose API which
requires significant changes to the interface to support.

We need to forward-port mongoose digest auth support as it is no
longer built into mongoose.

We also need to forward port multipart stream handling.

We need to overhaul the broadcast thread handling as the current
implementation is not thread safe and can drop messages.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>
Tested-by: Stefano Babic <sbabic@denx.de>