Introduce generic copy handler
The current "rawcopy" handler is quite limited. It just works as "raw"
handler. One common use case is to orvide a copy from a MTD device, and
then it is useless. Add a copy handler that runs in a chain another
handler, like "flash", so that any handler can be reused.

Signed-off-by: Stefano Babic <sbabic@denx.de>