Add option to link bootloader libraries statically
Commit 3a4e626b857 introduced dynamic linking for the bootloader
interface. This helps with standard Linux distros, where SWUpdate is
provided as package and it is not built from sources, and SWUpdate
checks at startup if the chosen library can be loaded. However, this
mechanism just works if dynamic linking is supported, and it fails on
system where dlopen() is not available, like uclibc and musl.

Add an option BOOTLOADER_STATIC_LINKED to reintroduce again static
linking.

Signed-off-by: Stefano Babic <sbabic@denx.de>