delta: force ZCK to not write temp files
ZCK library during zck write will write all (compressed) chunks into a
temporary file, and it will concatenate it to the generated zck header
at the end. SWUpdate needs just the header, and writing the chunks
requires a huge amonunt of memory. ZCK introduces a flag for solve this
to allow to create the header without saving the chunks, and SWUpdate
just sets the flag via zck_set_ioption().

Here the issue in zchunk:

	https://github.com/zchunk/zchunk/issues/89

Signed-off-by: Stefano Babic <sbabic@denx.de>