Wrong count of total steps
The count for steps to be run is wrong. Each script runs twice, even if
one run can be skipped if it is just a pre- or postinstall script. This
increases the number of steps to be run.

Note that partitions are already counted because they are just like
artifact that must be installed first.

Signed-off-by: Stefano Babic <sbabic@denx.de>