handlers: diskpart_handler.c: fix partition name
If a partitioned device's path ends in a digit, an interstitial 'p' is
added in the partition's device path.  For example, /dev/vdb has /dev/vdb1
but /dev/mmcblk0 has /dev/mmcblk0p1.
So instead of simply adding the partition number, we look for the
partition's name in /sys/block.

Signed-off-by: Philippe Reynes <philippe.reynes@softathome.com>