progress: do not block when message is sent
It is supposed that all clients registring themseleves with the progress
interface are reading for incoming events. However, bad clients can
block themselves for long time, and this can cause a deadlock. SWUpdate
will go on to send messages until the local socket is full, and then it
blocks waiting that a client unblock itself and go on to read. This will
block the update process, and SWUpdate will hang.

Change the behavior and sends events in not blocking mode. If it is
recognized that a client is not reading, the client is drop from the
list and no other messages are sent. Client must connect itself again.

Signed-off-by: Stefano Babic <sbabic@denx.de>