progress: add function to increase number of steps
It is not always possible to know the number of steps before updating.
There are some cases where a handler can split the update in multiple
single steps, calling a chained handler. A chained handler will update
the progress as ususal, but this means that a single step is multiplied
causing a total progress > 100%.

Add a funtion to increase during the update the number of steps.

Signed-off-by: Stefano Babic <sbabic@denx.de>