Expose `retrywait` setting to downloader
Even though the downloader already counts with a retry-sleep logic that
adds a delay between failed retries, that option is not exposed to the user
via arguments.

Currently 0s is the default delay between failed attempts. In my experience,
when dealing with unstable connections, it is desirable to have some delay
between retries, to increase the chance of having a successful connection.

This patch sets the default delay to default 5s, and also gives the possibility
of the user to customize it via downloader's args and settings, as it is already
the case for SUricatta.

Signed-off-by: João Loureiro <joaofl@gmail.com>