Remove a progress listener after some retries
Removing a listener after the first time the socket is full is too
harsh. There are conditions when SWUpdate is scheduled faster and
generates a lot of events, while the listeners are trying to read back
but not fast enough. Add a hard-coded retry mechanism, and try to send
an event again in case of EAGAIN or EWOULDBLOCK error. After some
retries, SWUpdate thinks that the listener is blocked and removes it
from the list.

Signed-off-by: Stefano Babic <sbabic@denx.de>