Allow to run a SWUpdate function in child process
Add a generic way for running an internal function in a child process.
This is helpful in case some external functions will be integrated and
it is easier to run them in a separate process than in core. The
function has the easy footprint:

	fn(int argc, char **argv)

The stdout and stderr are still replaced and output is sent to
SWUpdate's logs else TRACE and ERROR.

Signed-off-by: Stefano Babic <sbabic@denx.de>