suricatta/lua: Properly type integer values
Since Lua 5.3, the conversion of a number to a string adds a .0 suffix
to the result if it looks like an integer so that, e.g., 2.0 is printed
as such and not as 2 which was the case in Lua versions prior to 5.3
(see https://www.lua.org/manual/5.3/manual.html#8, Section 8.1).

Hence, fix this purely cosmetic glitch by explicitly using
lua_pushnumber() or lua_pushinteger() depending on the type.

Signed-off-by: Christian Storm <christian.storm@siemens.com>