network_thread: break out of the case on the notification history
There is a loop to send the notification history on the control
socket and if there was an error in that loop, then the code
would just continue with an unlocked mutex and progressing into
the control path. This error path seems to be very rare (if ever)
and so unlikely to be hit in any testing without some external
conditions shutting down sockets.

Signed-off-by: Charles Hardin <charles.hardin@chargepoint.com>