Fix calculation and reading of end-of-file padding
The stream interface may not read all data from the socket which could
cause swupdate to close the socket fd before a client has written all
data, resulting in an EPIPE on the client.

Instead of trying to calculate and expect a specific number of padding
bytes, perform best effort to read up to 512 bytes or until an error or
timeout is encountered.

Signed-off-by: Nick Potenski <nick.potenski@garmin.com>