bootloader: support for NVIDIA cboot - bootinfo
Some NVIDIA Tegra SOCs do not support U-Boot. NVIDIA has developed an
own bootloader as replacement for U-Boot. cboot (it should be patched)
can have access to the "bootinfo" database, that is stored on dependency
of the SOC type on fixed memory. The tegra-boot-tools provide a
convenient library to access programmatically to the environment. The
environment is redundant like in U-Boot, and this gurantees to have a
power-cut safe update of the variables.

This patch intoduce Tegra's "bootinfo" as bootloader interface, and it
depends on the libtegra-boot-tools library.

Signed-off-by: Stefano Babic <sbabic@denx.de>