swupdate_decrypt: add option to harden logging
For streamed AES-CBC encrypted images, an adversary can target each
16-byte block of encrypted data within an image and decrypt it, if they
can apply a huge amount of manipulated firmware updates and observe the
logged messages. On average, 2048 update attempts are needed for each
block.

For use cases where this scenario poses a risk, a config option is
introduced to suppress logging messages related to a hash mismatch and
errors in the decryption finalization (padding) of a streamed image.

Signed-off-by: Raimar Sandner <raimar.sandner@sick.de>