swupdate_decrypt: add error message for unsupported
Whether SWUpdate supports encrypted images or not is a compile-time
option. If SWUpdate was built without support for encrypted images and
an encrypted image is attempted to be installed, SWUpdate fails with an
generic error message:

[ERROR] : SWUPDATE failed [0] ERROR core/cpio_utils.c : __swupdate_copy : 498 : decrypt initialization failure, aborting

While this is certainly true, it helps little in finding the root cause.
This commit adds a more descriptive error message, leading directly to
the root cause.

Signed-off-by: Michael Adler <michael.adler@siemens.com>
Signed-off-by: Christian Storm <christian.storm@siemens.com>