fix(Makefile): correct argument count for trylink invocation
The trylink function expects to receive 7 arguments, with LDLIBS as the
last one. Previously, the Makefile was erroneously passing 8 arguments,
which included an empty string followed by LDLIBS. This commit rectifies
this off-by-one error by eliminating the unnecessary empty string,
ensuring that LDLIBS is accurately passed as the seventh argument.

Signed-off-by: Michael Adler <michael.adler@siemens.com>
Signed-off-by: Christian Storm <christian.storm@siemens.com>