ebg: ensure env_get returns valid strings
In the case of efibootguard, env_get might return a value which is not
null-terminated because this is allowed for user-defined variables.
However, SWUpdate assumes the returned value is a string, i.e. is
null-terminated. For example, the Lua function get_bootenv allows to
retrieve user-defined values.
This commit adds a check to ensure the returned value is indeed
null-terminated.

Signed-off-by: Michael Adler <michael.adler@siemens.com>
Signed-off-by: Christian Storm <christian.storm@siemens.com>