ebg: fix integer underflow
bgenv_get returns a negative value (-ENOENT) if the variable does not
exist, which results in an integer underflow if the variable `size`
is unsigned. Currently, this results in an attempt to malloc approx. 16
exabytes, which luckily fails on most systems.
The solution is to check for a valid (i.e. positive) size before trying
to malloc it.

Signed-off-by: Michael Adler <michael.adler@siemens.com>
Signed-off-by: Christian Storm <christian.storm@siemens.com>