handlers: ubivol: add rename fallback for replace property
The replace property currently requires the volume name to be replaced
to have already been created, however this creates a problem for cases
where we are doing an initial image installation but still want the
replace behavior to function normally when the volume name already
exists.

To handle these cases simply rename the volume to the replace target
instead of erroring out.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>
Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>