Add "vars" section in sw-description
The vars section will help to maintain a list of configuration variables
that can be set during an update, but are not relevant for the
bootloader. They are not stored in the bootloader's environment and
SWUpdate makes usage of libubootenv to store them with an own namespace.
This requires libubootenv 0.35 and higher.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>