Introduce SWUpdate persistent variables
The reason for this is that there are many cases where an update does
not require a reboot and does not require to involve the bootloader.
Using bootloader's environment (mainly used up now) as several
side-effects, for example for bootloader that foreseen states but not to
set an arbitrary variable to a value (like EBG for example), and there
is no need to change something for the bootloader.

However, in case the bootloader is not involved, but a reboot is still
required, SWUpdate needs to store persistently how the update ran. The
libubootenv library is the interface to U-Boot, but it is able to store
a U-Boot like environment (variable=value) on different places. Release
0.3.5 is able to handle multiple environment by extending the
configuration file, and this can be used to have an environment for
bootloader and an environment for SWUpdate.

Not that libubootenv does not require U-Boot and can be used on systems
with other bootloaders.

The patch here store the variables set in sw-descriprion via the "vars"
section in a libubootenv namespace configured at start up.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>