BUG: JSON parser can't handle a string "offset"
get_field(p, elem, "offset", &offset) in case of JSON does not return 0
if field is a string, causing a wrong offset to be evaluated. This bug
does not happen if libconfig is the parser.

Call is_field_numeric() in advance to check if offset is set as number
or string.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>
Reported-by: Einar Jón <tolvupostur@gmail.com>