Store globally namespace for swupdate_vars
Namespace (see also libubootenv) for SWUpdate's own variable is defined
in configuration file. At the moment, SWUpdate will support to write
just into a single namespace. Most processes do not need to know it, and
it is ok to define a default mechanism to retrieve the namespace. This
patch adds a singleton for namespace.

AWupdate will store the selected namespace at the startup and will use
it in calls where the namespace is not set.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>