Implement IPC to set / get SWUpdate's vars
Access to SWUPdate's vars (namespace) is guaranteed for the core, but
other processes could need to store some values to store a state between
runs. This is not possible from subprocess because rights are not
guaranteed for the namespace's storage.

Add an IPC as done for bootloader environment to let subprocesses to
store persistently variables.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>