Enum for cause field in progress messages
The cause field is unused up now and just zeroed. Add an enum to support
info of specific type and replace occurencies where zero is forced.

The first introduced cause is CAUSE_REBOOT_MODE to infor if the update
requires a reboot or it is a on the fly update.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>