Send information to backend in case of on the fly updates
Not all updates require a reboot. In many cases, some parts of the
system are updated and a reboot is not necessary (on the fly update).
However, communication with the backend stops because software was
already installed successfully and to close the transaction, SWUpdate
had to be restarted.

To better support on on-the-fly updates, a new attribute "no-reboot" is
added. This is just to inform all processes registered with the progress
interface that a reboot at the end of the update is not necessary.
SWUpdate registers this attribute, and sends via IPC a message to the
suricatta module (currently, only Hawkbit is supported) to inform that
the update was successfully executed and the transaction can be closed.

After that, the suricatta module exits from the wait process and it is
able to receive further updates, too.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>