ucfw: Refactored gpio handling
This commit splits the switch_mode function into the
gpio-initialization, the actual mode switch and the cleanup phases. This
is mainly done in order to prepare for adding gpiolib v2 api capability
but has the added advantage, that the gpios are no longer released
during the update, which for one prevents the gpios from floating, and
for another, stops other processes from taking over the pins during an
update phase.

Signed-off-by: Boerge Struempfel <boerge.struempfel@gmail.com>