fix: properly cleanup sockets on SIGTERM
When running with a non default TMPDIR, the IPC sockets were not cleaned
up correctly. The reason for that was an incorrect detection if swupdate
is started as systemd service. In addition, the creation logic and
cleanup logic was spread across multiple files and non consistent.

This is now fixed by tracking which sockets were created (this can variy
if sockets are passed by systemd) and deallocating these in the atexit
handler. For that, a new network_utils unit is added which provides
functions to globally register sockets that should be cleaned up.
By that, duplicate logic across swupdate is also removed.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Acked-by: Stefano Babic <stefano.babic@swupdate.org>