progress: versioning of the API
It is not excluded that the message structure use to exchange messages
via progress interface can be changed in future, raising compatibility
issues. To check this, the unused "magic" field assumes the value for
the current API version. Schema for the versioning is put into
include/progress_ipc.h. The receiver (progress_ipc_receive) is changed
to check for changes in the major version of the API, returning an error
in that case.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>