channel_curl: add fifo to read callback
In case of upload, the data is read from a buffer that must be allocated
and filled by the caller (request_body field in channel_data_t). This is
not suitable in case a large amount of data should be upload, for
example in case a big image should be transferred.

Let the callback to read from a fifo that must be created by the caller.
The caller writes then into the FIFO, and the callback consumes the data
by reading from the FIFO.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>