Lua: prepare to extend Lua lifetime
A Lua context is created and destroyed for each Lua scriptthat should
run. This isolates each script, but it does not allow that a later
script can benefit and load a function previously defined.

It was decided that so high isolation is not necessary - instead of
having a context for each script, create a Lua context for each
installation request. The Lua context is then initialized when
sw-description is parsed and removed when the installation has finished,
independently from the state.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>