BUG: post-update command is no longer called from mongoose interface
post-update is not called anymore. Reason are commit 8fccd13 and d116870 because
they force the source of an update independently from the real source.
The source for an update is set when a install request is sent and cached
into core. The progress interface must retrieve the source and be stateless
for it.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>
Reported-by: Michael Glembotzki <Michael.Glembotzki@iris-sensing.com>