Lua Script: runs with global Lua state
Add a property "global-state" to run the script with a per installation
Lua state or creating a new local as in the past.

If global-state = "true" (as property, value is always a string), Lua
state is reused and the script can call any Lua function that was loaded
before during the installation.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>