Lua Script Handler: allows to run scripts globally
Up now, scripts have run in isolated context: SWUpdate allocates a new
Lua state, and import the basic libraries before loading the script. A
script is then isolated, but it cannot access to function already
loaded, or it is not possible to reuse functions from 2 or more scripts.

With the introduction of a per installation Lua state, Lua scripts can
call functions already defined in previous scripts, or defined in
sw-description. Because when a script is loaded, it could overwrite
existing functions with the same name, it was decided that functions in
scripts must be unique, that is each function should be declared just
once during an installation process.

This means that for global state,  sw-description should contain the
name of the function for each step (pre- , postinstall or postfailure)
that should be called: the names preinst, postinst and postfailure are
still valid in case the script runs with isolated state.

This allows also to load a script without executing if no functions are
defined, and functions in the script can be called by later scripts.

Note that the handler will load the script in case of global state just
once during the "preinstall" call. Later, it is assumed that functions
will be already available.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>