swuforward: add generic uploader
The handler is used to push a SWU to other systems running SWUpdate.
This patch introduce the concept of generic uploader, and the handler
can upload an artifact (not only a SWU) to a URL. The handler will still
connect via Websocket to the webserver, and it will parse the answer. To
make it generic, the answer can be parsed by a custom Lua function, that
should be loaded previously via embedded-script or via a script, or even
via swupdate_handlers.lua at the startup.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>