network_thread: block SIGPIPE in network thread
SIGPIPE was blocked in the subprocess thread since commit a781be1d0b7a
("network_thread: Don't SIGPIPE-die on IPC client crash"), but the main
thread sometimes sends reply through write() directly (for e.g. STATUS)

In this case, swupdate crashes if the client is gone, which is not
something we want.

Since we block SIGPIPE before creating the subprocess thread we no
longer need to block it there.

(Add errno in write error message while we are here)

Signed-off-by: Shiita ISHIGAKI <shiita.ishigaki@atmark-techno.com>
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>
Acked-by: Stefano Babic <stefano.babic@swupdate.org>