suricatta process notification: improve ipc_get_status scheduling
The process notification thread gets messages from swupdate to submit
them to the hawkbit server.

Before this patch, ipc_get_status_timeout() was called in a tight loop,
probably with the assumption that it would wait until the timeout
happens if no message was available.
Unfortunately, the server just replies immediately that there is no
message, so this would just hammer the server in a tight loop.

Furthermore, on slow devices if the server takes more than 100ms to
reply we close our socket so the server would crash to sigpipe without
the previous fix.

Adjust this loop to:
- since ipc_get_status() always returns immediately the timeout is
useless, use the version without timeout.
- wait 100ms before calling ipc_get_status() again if no message was
available.

Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>
Acked-by: Stefano Babic <stefano.babic@swupdate.org>