hawkbit: check for validity of action_id
The "swupdate-ipc sendtohawkbit" tool allows to send a feedback to the
Hawkbit server with external data. Commit d770ad59e introduces that
SWUpdate can persistently store the action-id between runs, and this
worked after a reboot when SWUpdate is started again.

This patch introduces the same feature when IPC is used. The action-id
in IPC can be set to 0 or -1 (invalid for Hawkbit), and SWUpdate will
then check if a valid action_id is in the environment.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>