Enforce type check in sw-description
Attributes in sw-description are strictly typed and SWUpdate expects to
get values for the defined type. With the exception of strings,
attributes are parsed using a generic get_field() function, that
retrieves the value in the format set into sw-description, that can
differ from the expected one.

Add a set of macros to be used and to replace the generic get_field,
that is GET_FIELD_<type>, like GET_FIELD_STRING that already exists, and
checks for type before retrieving the value.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>