suricatta: prefer CLOCK_MONOTONIC for suricatta_wait
When calling sem_timedwait there may be side effects caused by
discontinuous jumps in the system time which can cause the clock
to go backwards.

To try and avoid side effects caused by these issues use the new
sem_clockwait function with CLOCK_MONOTONIC which should be available
on linux as of glibc version 2.30. Other libc's do not appear to have
support for sem_clockwait so we still have to fall back to
sem_timedwait on those systems.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>