cpio_utils: fill_buffer: fix return value on EOF after short read
fill_buffer would return 0 after having filled buffer/updated checksums
if it reaches the end of file (returns 0) after another read.

In practice, this never happens because the caller knows how much to
read, and chunks are smaller than what had been previously written in
the pipe so this is a theorical problem, but because the checksums are
correctly updated until the end this would not be detected if it happens
in practice (the end of the data would just not be fed to the handler),
so this is best fixed.

Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>
Acked-by: Stefano Babic <stefano.babic@swupdate.org>