hawkbit: do not reset action_id when getting from env
server_handle_initial_state() firsts get an action id from the server
and then overrides it with the env value, but there is no check if there
had been no env value, sending a request to id -1 to server.

The commit that added this reset also added code to only call
get_action_id_from_env() if action_id <= 0, so it did not need to reset
it: just leave the status quo of using the last value if it wasn't set.

Fixes: f3b1e1e26dfb ("hawkbit: check for validity of action_id")
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>
Acked-by: Stefano Babic <stefano.babic@swupdate.org>