handler: ubivol: relax accessing UBI volume
It could be that some race conditions happens when SWUpdate tries to
lock the UBI volume for updating, for example after creating the volume
if udev is running and it accesses for some milliseconds. This is not an
issue inside SWUpdate, but SWUpdate can retry to get exclusive access of
the volume before giving up, making the update itself more robust.

Signed-off-by: Stefano Babic <stefano.babic@swupdate.org>
Reported-by: Christian Eggers <ceggers@arri.de>
Tested-by: Christian Eggers <ceggers@arri.de>